% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longest_word.R
\name{longest_word}
\alias{longest_word}
\title{The Longest Word}
\usage{
longest_word(text)
}
\arguments{
\item{text}{A character string containing the text to analyze.}
}
\value{
A data frame with the longest word, its position and length.
}
\description{
This function identifies the longest word, its position and its length in input the text.
}
\examples{
#How to use longest_word()
text <- "A sentence with words of varying lengths as an example"
longest_word(text)
}
