% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence_summary.R
\name{sentence_summary}
\alias{sentence_summary}
\title{Sentence Summary}
\usage{
sentence_summary(text)
}
\arguments{
\item{text}{character string containing the text to analyze. The text should
be composed of sentences separated by periods (.), exclamation marks (!), or
question marks (?).}
}
\value{
A list containing the sentence count, average word count, longest word, shortest word and readability score.
}
\description{
This function summarizes each sentence of the input text
by counting the sentences, calculating the average sentence length,
calculating the readability score and giving you the longest and shortest sentences of your input.
}
\examples{
#How to use sentence_summary()
text <-  "This is an example. It has more than two sentences! Well, this is the third."
sentence_summary(text)
}
