% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.analyzer.R
\docType{package}
\name{text.analyzer}
\alias{text.analyzer-package}
\alias{text.analyzer}
\title{text.analyzer Package}
\description{
The \strong{text.analyzer} package provides simple tools to analyze text data quickly and effectively in R.
}
\details{
It includes functions to:
\itemize{
\item Summarize basic text characteristics (\code{text_summary})
\item Identify the longest word and its position (\code{longest_word})
\item Count the 10 most frequent words, excluding common stop words (\code{top10_words})
\item Analyze sentence structure and readability (\code{sentence_summary})
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YusraKunduzi/text.analyzer}
  \item Report bugs at \url{https://github.com/YusraKunduzi/text.analyzer/issues}
}

}
\author{
\strong{Maintainer}: Yusra Kunduzi \email{yusra.kunduzi@student.hu.nl}

}
