% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top10_words.R
\name{top10_words}
\alias{top10_words}
\title{The 10 Most Frequent Words}
\usage{
top10_words(text)
}
\arguments{
\item{text}{A character string containing the text to analyze.}
}
\value{
A vector of the top 10 most frequent words and their counts.
}
\description{
This function counts the 10 most frequent words in the input text,
excluding common words like "the" in English and Dutch.
}
\examples{
#How to use top10_words()
text <- "A character string containing example text.
The cat sat on the mat, and the cat looked at the dog while the dog barked at the cat"
top10_words(text)

}
